[OS-36] WIP: Implement User-Workspaces relation
