+++ server/public/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 
 <head>
-  <title>server</title>
+  <title>Welcome to Onus server</title>
 
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
