+++ server/src/repositories/index.ts
@@ -1,4 +1,4 @@
-export * from './users.repository';
+export * from './user.repository';
 export * from './workspace.repository';
 export * from './status.repository';
 export * from './task.repository';
