+++ server/src/services/user-service.ts
@@ -4,15 +4,15 @@ import { repository } from '@loopback/repository';
 import { HttpErrors } from '@loopback/rest';
 import { securityId, UserProfile } from '@loopback/security';
 import { PasswordHasherBindings } from '../keys';
-import { UsersRepository, Credentials } from '../repositories';
+import { UserRepository, Credentials } from '../repositories';
 import { User } from '../models';
 
 import { BcryptHasher } from './hash.password';
 
 export class MyUserService implements UserService<User, Credentials>{
   constructor(
-    @repository(UsersRepository)
-    public userRepository: UsersRepository,
+    @repository(UserRepository)
+    public userRepository: UserRepository,
 
     @inject(PasswordHasherBindings.PASSWORD_HASHER)
     public hasher: BcryptHasher
